requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
connexion==2.7.0
Flask==1.1.2
idna==2.10
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
openapi-spec-validator==0.2.9
protobuf==3.14.0
pyrsistent==0.17.3
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
swagger-ui-bundle==0.0.8
urllib3==1.26.2
Werkzeug==1.0.1
